Move the paginator termination tests into tests/test_helpers.py
Per review: they belong with the other helper tests rather than in their own
module, and the @pytest.mark.asyncio decorators are unnecessary because
setup.cfg sets asyncio_mode = auto.

The local helper is renamed _paginator_pages now that it shares a module with
the rest of the helper tests. Behaviour is unchanged: reverting the source fix
still fails both tests, and the suite is 194 passed.
